net: wireless: sd8977: Allow BA of TID 6,7
The problem is that NETGEAR AP(access point) keep trying to send ADDBA
request for TID 6 even it was rejected lots of times already. This
NETGEAR AP would stop sending ADDBA request for TID 6 when it receives a
ADDBA response with success.
This patch allows ADDBA for TID 6 and 7 to resolve this problem. But
this patch can cause voice problem when we are trying to use voice
related applications in noise environment.

Change-Id: I7d1ab2944283440e088da067b178b845dce56456
Signed-off-by: Soon-Yong Cho <soonyong.cho@samsung.com>