clk: introduce nxp3220 clock driver
The nxp3220 clock controller generates clocks to various peripherals of nxp3220
SoC. It is consist of multiple PLLs and CMUs.

* List of PLLs
- PLL0
- PLL1
- PLL_CPU
- PLL_DDR0
- PLL_DDR1

* List of CMUs
- CMU_SRC: main clock sources across CMU_SYS/CPU/DDR/USB/MM
- CMU_SYS: system clocks
- CMU_CPU: cpu clocks
- CMU_DDR: ddr memory clocks
- CMU_USB: usb clocks
- CMU_MM: multimedia clocks

The CMU_SRC can be main clock sources which include MUX and Dividers
across the other clock domains(SYS/CPU/DDR/USB/MM). Actually, the
CMU_SRC should be implemented as separate driver apart from CMU_SYS but
this driver just implemented them to nxp3220-cmu one driver. If not to
do that, we have to specify tons of input clocks for CMU_SYS.

Change-Id: Ie9ef737b7723acfd105d5ae1ff855926cefe7142
Signed-off-by: Chanho Park <chanho61.park@samsung.com>