clk: nexell: convert to nxp3220 gate clock
This patch converts the general gate clock to the nxp3220 gate clock.
To do that, I added a clr_offset field which will be used from the
nxp3220 gate clock. After this patch, unused clocks will be
automatically turned off end of boot stage. To prevent turning off the
unused clock, we can add the CLK_IGNORE_UNUSED flag if the clock should
not be turned off.

Change-Id: I641194c2ab61ff118ad46252f357f9cc35a16725
Signed-off-by: Chanho Park <chanho61.park@samsung.com>