arm: dts: artik310s: optimize frequency and strength of SD
This patch modifies clock-frequency of SD. The SD frequency of
artik310s is 50MHz and DWMCI divider in artik310s is 4. We can use
both of 200MHz and 400MHz, so choose 200MHz due to coherence between
u-boot and kernel.

This patch modifies strength of SD clock from 0 to 1.
There sometimes was SD failures due to strength. This patch
fixes these problems.

Change-Id: Ib4cab50a6e9dbf2fca6aac51d78631ba1ebf182a
Signed-off-by: Sowon Na <sowon.na@samsung.com>