arm: dts: artik310s-explorer: enable drm/hdmi as dtbo
This patch enables a RGB overlay node for artik310s explorer board. To
be enabled the rgb and hdmi node of the board, i2c2/i2c3/i2c4/mac1
should be disabled because the pins of rgb are conflicted with them.
Due to restriction of sii9022 hdmi converter chip, the max clock speed
of the DDC line should be limited to 50KHz.

Change-Id: I7a17096f4e19542ef0c20da30238a59defca52c1
Signed-off-by: Chanho Park <chanho61.park@samsung.com>