arm: dts: artik310s: add keep-power-in-suspend for sd8977 driver
The Marvell sd8977 driver prevents the system to enter into a suspend state if
the card power is not preserved during a suspend/resume cycle.
This patch adds the keep-power-in-suspend power management property to the
SDIO/MMC node so the sd8977 suspend handler doesn't fail and the system is able
to enter into a suspend state.

Change-Id: Ib12cb2aa8fbf02256fbd03898fa983ee4d80b0d3
Signed-off-by: Soon-Yong Cho <soonyong.cho@samsung.com>