clk: nexell: set CLK_SET_RATE_NO_REPARENT of dpc source clock
This patch fixes the dpc source clock can be attached to different pll
sources such as pll_cpu. The frequency of cpu's pll can be changed
frequently but the dpc clock cannot be changed until we set a notifier
of clock change. Instead of setting the notifier, we just set
CLK_SET_RATE_NO_REPARENT flag of the dpc source clock.

Change-Id: Ia24a512e3af41773c6982a9a6d896ecc21cbdc90
Signed-off-by: Chanho Park <chanho61.park@samsung.com>