arm: dts: nxp3220: use macro for describing retention pins
This patch changes gpio retention pin descriptions through macro in
order to improve readability and easy editing. Previous style was good
to define a retention group at once but it can be really hard to
override or edit by child DT file. So, instead of defining several pins
as a group, this patch defines a retention group per a pin.

Change-Id: I8be8802f7818f14360bd237e6673a99f8adb0656
Signed-off-by: Chanho Park <chanho61.park@samsung.com>