dts: nxp3220 evb: Modify the contents of the device node for spi boot mode
This patch modifies the contents of the device node for spi boot mode.
To fix the problem that is not booted up after rebooting by setting
"4-byte-address mode" after spi boot, add "disable-4byte-address"
so that 4byte address mode is not used when booting spi, mode.
It also changes the partition of the data to preserve the data for the
spi boot.

Signed-off-by: JongKeun Choi <jkchoi@nexell.co.kr>
Change-Id: Idbb09b8fd12898c8eb480ae00a32f2bf9dee61de
Reviewed-on: https://review.gerrithub.io/c/NexellCorp/linux_kernel-4.14.x/+/449387
Reviewed-by: jung hyun kim <jhkim@nexell.co.kr>
Tested-by: jung hyun kim <jhkim@nexell.co.kr>