arm: dts: nxp3220: Fix m_can0/1 core clock rate to 100Mhz
This patch Fix m_can0/1 core clock rate to 100Mhz

M_CAN datasheet:
In order to achieve a stable function of the M_CAN, the Host clock (hclk)
must always be faster than or equal to the CAN clock (cclk).
Also the modulation depth of a spread spectrum clock has to be regarded.

Signed-off-by: jhkim <jhkim@nexell.co.kr>
Change-Id: Id6568097e901cf8a65a535ba993dfb959a01e2d1
Reviewed-on: https://review.gerrithub.io/c/NexellCorp/linux_kernel-4.14.x/+/467817